requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22